minor fixes

+++ commands/setlanguage.js
@@ -15,16 +15,16 @@ module.exports = {
 		else {
 			const newLanguage = args[0].toLowerCase(),
 				ref = db.collection('servidores').doc(message.guild.id);
-			const oldLanguage = ref.get('language') || config.language;
+			const oldLanguage = ref.get('language');
 
-			if (oldLanguage == newLanguage) {
-				return message.reply(lang.error.sameLanguage).catch();
-			}
-			else if (newLanguage == config.language) {
+			if (newLanguage == config.language && newLanguage != oldLanguage) {
 				languages[message.guild.id] = config.language;
 				ref.update({
 					language: FieldValue.delete(),
 				}).catch(err => { console.error(err); });
+			}
+			else if (newLanguage == config.language || newLanguage == oldLanguage) {
+				return message.reply(lang.error.sameLanguage).catch();
 			}
 			else {
 				languages[message.guild.id] = newLanguage;
